backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.0",
    "altair>=5.5.0",
    "fastapi[standard]>=0.115.4",
    "google-cloud-bigquery>=3.27.0",
    "pandas>=2.2.3",
    "python-dateutil>=2.9.0.post0",
    "sqlalchemy>=2.0.36",
]

[dependency-groups]
dev = [
    "coverage>=7.6.4",
    "mypy>=1.13.0",
    "pre-commit>=4.0.1",
    "pytest>=8.3.3",
    "ruff>=0.7.1",
    "sqlfluff>=3.2.5",
]
repl = [
    "ipykernel>=6.29.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
strict = true
exclude = ["venv", ".venv"]

[tool.ruff]
target-version = "py312"
